requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.8.0
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.13.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.6.0
apispec==6.3.0
apturl==0.5.2
argcomplete==3.2.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brlapi==0.8.3
cachelib==0.9.0
cachetools==5.3.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
cupshelpers==1.0
db-dtypes==1.2.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
duplicity==0.8.21
email-validator==1.3.1
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.0.0
gcloud-aio-storage==9.0.0
gcsfs==2023.12.2.post1
google-ads==22.1.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.38.1
google-cloud-appengine-logging==1.4.0
google-cloud-audit-log==0.2.5
google-cloud-automl==2.12.0
google-cloud-batch==0.17.6
google-cloud-bigquery==3.14.1
google-cloud-bigquery-datatransfer==3.13.0
google-cloud-bigquery-storage==2.24.0
google-cloud-bigtable==2.22.0
google-cloud-build==3.22.0
google-cloud-compute==1.15.0
google-cloud-container==2.36.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.17.0
google-cloud-dataflow-client==0.8.6
google-cloud-dataform==0.5.5
google-cloud-dataplex==1.10.0
google-cloud-dataproc==5.8.0
google-cloud-dataproc-metastore==1.14.0
google-cloud-dlp==3.14.0
google-cloud-kms==2.20.0
google-cloud-language==2.12.0
google-cloud-logging==3.9.0
google-cloud-memcache==1.8.0
google-cloud-monitoring==2.18.0
google-cloud-orchestration-airflow==1.10.0
google-cloud-os-login==2.12.0
google-cloud-pubsub==2.19.0
google-cloud-redis==2.14.0
google-cloud-resource-manager==1.11.0
google-cloud-run==0.10.1
google-cloud-secret-manager==2.17.0
google-cloud-spanner==3.40.1
google-cloud-speech==2.23.0
google-cloud-storage==2.14.0
google-cloud-storage-transfer==1.10.0
google-cloud-tasks==2.15.0
google-cloud-texttospeech==2.15.0
google-cloud-translate==3.13.0
google-cloud-videointelligence==2.12.0
google-cloud-vision==3.5.0
google-cloud-workflows==1.13.0
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
greenlet==3.0.2
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-gcp==0.2.2
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
hdfs==2.7.3
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
json-merge-patch==0.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keyring==23.5.0
kubernetes==28.1.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.20.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.2
olefile==0.46
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pandas-gbq==0.20.0
paramiko==2.9.3
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
pluggy==1.3.0
prison==0.2.1
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydata-google-auth==1.8.2
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.32.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
rsa==4.9
SecretStorage==3.3.1
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
sqlalchemy-bigquery==1.9.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.6.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
ssh-import-id==5.11
statsd==4.0.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
universal-pathlib==0.1.4
uritemplate==4.1.1
urllib3==1.26.18
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
websocket-client==1.7.0
Werkzeug==2.2.3
wrapt==1.16.0
WTForms==3.1.1
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==3.17.0
